requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
celery
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
minio
PyPDF2
python-docx
python-dotenv
pydantic-settings
